+++ RELEASENOTES.md
@@ -2,7 +2,8 @@
 
 ### 3.5.1 (UNRELEASED)
 * Added the ability to disable BMLT based authentication in favor of local auth [#339]
-* Several reports fixes [#342] [#343] [#344] [#345]
+* Added service body IDs in dropdowns on the admin UI. [#349]
+* Several reports fixes [#342] [#343] [#344] [#345] [#346] [#348]
 * Fix for Play link not working on Call Handling modal for service bodies [#347]
 
 ### 3.5.0 (December 4, 2019)